backend/requirements.txt
fastapi
langchain
openai
bs4
requests
pydantic
python-dotenv
langchain_community
langchain_openai
google-search-results
faiss-cpu
gunicorn
uvicorn[standard]
pytest
black
brotli
aiohttp